Draw tool now draws.  Need to refine these tools.  Also need configuration system implemented for the tools.  TODO: create a Tool Properties Dialog.
git-svn-id: https://avogadro.svn.sourceforge.net/svnroot/avogadro/trunk@83 af78ad9f-9711-0410-ba57-a8eff4914ec4

+++ avogadro/src/tools/selectrotate.cpp
@@ -165,7 +165,7 @@ void SelectRotate::mouseMove(GLWidget *widget, const QMouseEvent *event)
   widget->updateGL();
 }
 
-void SelectRotate::selectionBox(int sx, int sy, int ex, int ey)
+void SelectRotate::selectionBox(float sx, float sy, float ex, float ey)
 {
   if(!_selectionDL)
   {
